// src/app/interface/datacalculate.interface.ts
export default interface DataCalculate {
   
   vlGoldPrice: number;
   trm: number;
   vlCopOZ: number;
   ozTyGr: number;
   vlGrGold: number;
   medalGr: number;
   materialVl: number;
   dateCurrent: any;
   utility: number;
   base: number;
   ivaValueConst: number;
   ivaVl: number;
   incrementIva: number;
   medalVl: number;
   vlMedalIncrement: number;
   increment: number;
}